Add the HHubGlobalBan plugin.
HGB is a plugin which adds an automatic, Global Ban system to your server. Have you heard of CAX Bans? This is it, but Community-Edition™️.
Malicious individuals such as cheaters, minge-groups, etc are added to this list and get banned from each server that has this plugin installed. The ban list is managed by me, and only I can add onto the list, so it is not exploitable in any way.
In the future I plan on adding a whitelist system, where server owners can add specific steamids to a list and prevent them from being banned on their server, should they be in the global ban list.
The plugin also features a 'Version Control' system. If there is ever a new version of the plugin released, you will see a message on your server console when you start up your server.