Updated weight
Made some updates to my weight plugin. Decided it was better to completely recalculate a character's carry weight everytime an update is needed instead of just modifying the existing carry value. In the long run this prevents problems where it wasn't being updated correctly. Checks were also added for vendors so if they can't carry an item they are attempting to buy they won't be able to complete the purchase.